[tests] New, refactor
